pyproject.toml
[tool.poetry]
name = "a-data-pro-internship"
version = "0.1.0"
description = ""
authors = ["Stoyan Ivanov <ssivanov19@codingburgas.bg>"]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.5.0"
pyodbc = "^4.0.31"
python-dotenv = "^0.18.0"
python-logging-loki = "^0.3.1"
sentry-sdk = "^1.3.0"
Django = "3.0.6"
python-decouple = "^3.4"
django-mssql-backend = "^2.8.1"
six = "^1.16.0"
flake8 = "^3.9.2"
black = "^21.7b0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
